Add a BGAppRefreshTask which allows the app to be restarted by the system
This might improve background task scheduling in certain cases when background
processing tasks are not being re-scheduled (for example after device reboot).
The SDK does not perform a sync in this case, as it might be too time
consuming for app refresh tasks (which get a max runtime of 30s).